pytorch-numba-extension-jit/pyproject.toml
[project]
name = "pytorch-numba-extension-jit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "p-adema", email = "117938385+p-adema@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "numba-cuda>=0.9.0",
    "torch>=2.7.0",
    "nvidia-cuda-runtime-cu12>=12.8.57; sys_platform != 'darwin'",
    "uuid>=1.30",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.27.0",
    "ruff>=0.11.7",
]

[[tool.uv.index]]
name = "torch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.uv.sources]
torch = [
    { index = "torch-cu128", marker = "sys_platform != 'darwin'" },
]
